requirements.txt
django-bootstrap3
pillow